Move remaining post action components into dropdowns/posts
