--- packages/lesswrong/components/posts/PostActions/index.ts
@@ -1,3 +1,0 @@
-import { importComponent } from '../../../lib/vulcan-lib';
-
-importComponent("PostActionsButton", () => require('./PostActionsButton'));
